pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autofillgluon"
version = "0.1.0"
authors = [
    {name = "Deniz Akdemir", email = "your.email@example.com"},
]
description = "Machine learning-based missing data imputation using AutoGluon"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Information Analysis",
]
dependencies = [
    "autogluon.tabular>=0.8.0",
    "autogluon.core>=0.8.0",
    "pandas>=1.5.0",
    "numpy>=1.20.0",
    "scikit-learn>=1.0.0",
    "lifelines>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0.0",
    "pytest-cov>=2.10.0",
    "black>=20.8b1",
    "isort>=5.0.0",
    "flake8>=3.8.0",
]
viz = [
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/AutoFillGluon"
"Bug Tracker" = "https://github.com/yourusername/AutoFillGluon/issues"
"Documentation" = "https://github.com/yourusername/AutoFillGluon#readme"